refactor: InputHandler 클래스 삭제
- InputHandler 클래스를 InputView 클래스로 통합
- RancingGameController에서 InputHandler 대신 InputView를 통해서 입력값 받기
